[REFACTOR] TrashDetector 기능별 컴포넌트 분리 및 리팩토링 #44
