+++ Assets/Scripts/SLAM/CameraCapture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae0f3971334363647bbd843f0ac23dc5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
