+++ Assets/Scripts/SLAM/TrashDetectorController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 601c829418d7b7f4c870c4dba368b990
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
